Completely Fixed Runline
Restructured Runline completely. Now it just use the ~ for a flag
always I think. I think this is more efficient and it has less of an
overhead. Also tweaked compressor a little bit. I think I am going to
put some sort of menu system in the Compressor class… but not until I
get LZ77 working. I also have to get LZ77 working with Huffman… I will
have to make a new version of Huffman that is slightly different.